refactor: move utility functions out of frappe/__init__.py
Move scrub/unscrub → utils.data, get_module/read_file/get_file_json/get_file_items/get_attr → utils, get_module_path/get_app_path/get_app_source_path/get_site_path/get_pymodule_path/get_module_list → modules.utils, get_all_apps/get_installed_apps → apps, respond_as_web_page/redirect_to_message → utils.response. All remain accessible as frappe.X via lazy imports.
